ai/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.5
python-dotenv==1.0.1
